refactor(lib): group lib files into api, auth, payload, wallet folders
